requirements.dev.txt
google-cloud-bigquery
numpy
pandas
seaborn
bokeh
argparse
openpyxl
pygal
cairosvg
spacy
